feat(apps/chat-with-pdf): switch document-switcher to document-title
