+++ apps/chat-with-pdf/app/components/header/document-title/constants/input-names.ts
