+++ apps/chat-with-pdf/app/components/header/document-title/new-document-dialog/new-document-dialog.tsx
